test: move tests under test/unit/ and the type reference under test/helpers/
wb resolves unit tests from `test/unit/`, and wb 19.1.0 makes the layout a hard
error in `wb lint` and `wb verify`. The `require('../temp/...')` calls are
file-relative and needed the extra level; the `path.resolve('test-fixtures', …)`
lookups resolve from the working directory and are unaffected.
