+++ test/unit/fileWithHyphen.test.ts
@@ -8,7 +8,7 @@ test('filename with hyphen', () => {
   const outFile = path.resolve('temp', 'fileWithHyphenI18n.ts');
   child_process.spawnSync('yarn', ['start', '-i', inputDir, '-o', outFile, '-d', 'en-us', '--global']);
 
-  const { changeLanguageByCode, i18n } = require('../temp/fileWithHyphenI18n.ts');
+  const { changeLanguageByCode, i18n } = require('../../temp/fileWithHyphenI18n.ts');
 
   expect(i18n.color()).toBe('color');
